requirements.txt
# UP Prepaid Engine Automation Framework
# Python dependencies

# Data manipulation and analysis
pandas>=1.3.0
numpy>=1.20.0

# HTTP requests and API calls
requests>=2.26.0

# PostgreSQL database connectivity
psycopg2-binary>=2.9.0

# Excel file handling
openpyxl>=3.0.0

# Optional: For API wrapper example
# flask>=2.0.0

# Optional: For enhanced logging
# colorlog>=6.0.0
